third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-with-pseudo-element-without-renderer.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
#element:after { display: none; color: green; }
</style>
<div id="element"></div>

<script>
test(function() {
    var computedStyle = getComputedStyle(document.getElementById('element'), ":after");
    assert_equals(computedStyle.color, 'rgb(0, 128, 0)');
    assert_equals(computedStyle.display, 'none');
}, 'getComputedStyle works on pseudo-element without renderer');
</script>
